requirements.txt
discord.py>=2.3.2
yt-dlp>=2023.11.16
PyNaCl>=1.5.0
python-dotenv==1.0.0
aiohttp>=3.9.1
yarl>=1.9.3
typing-extensions>=4.8.0
pyyaml>=6.0.1
asyncio>=3.4.3
uvloop>=0.17.0; sys_platform != "win32"  # Windows 외 시스템에서 성능 향상